backend/requirements.txt
channels
django
django-cors-headers
django-rest-framework
djangorestframework-simplejwt
uvicorn
daphne